refactor(tui): extract command-action dispatch from app.rs run loop
app/app.rs was 1200 lines, dominated by the run() loop. Move the
~260-line `match resp.action { CommandAction::… }` block into
app/command_action.rs::handle_command_action, which the loop now calls.
app.rs drops to 945 lines. No behavior change.
